+++ Requirements/Views/RequirementsBrowser.xaml
@@ -31,7 +31,7 @@
             <cdp4Composition:ThingToIconUriConverter x:Key="ThingToIconUriConverter" />
             <converters:RequirementStateOfComplianceColorConverter x:Key="RequirementStateOfComplianceColorConverter"></converters:RequirementStateOfComplianceColorConverter>
             <converters:RequirementStateOfComplianceToVisibilityConverter x:Key="RequirementStateOfComplianceToVisibilityConverter"></converters:RequirementStateOfComplianceToVisibilityConverter>
-            <selectors:RequirementsTreeListNodeSelector x:Key="RequirementsTreeListNodeSelector" />
+            <selectors:RequirementsTreeListNodeImageSelector x:Key="RequirementsTreeListNodeSelector" />
 
             <Style TargetType ="{x:Type dxg:RowControl}" x:Key="RowStyleDeprecatedRequirement">
                 <Setter Property="ToolTip" Value="{Binding Row.Tooltip}" />
